Generate homepage from pad metadata
`python manage.py index` scrapes the site for RDF/A and than generate an index from it

What’s nice is that this part doesn’t need a database—
we create the graph each time from scratch.

We then convert the RDF/A to a form the template can handle, store it in a json,
and the index view uses this.